backend/requirements.txt
base58==2.1.1
celery==5.5.3
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
py_near==1.1.59
python-dotenv==1.1.1
python_binance==1.0.30
Requests==2.32.5
solana==0.36.9
solders==0.26.0
tenacity==9.1.2
tonsdk==1.0.15
web3==7.14.0
